Organize vimrc file into folded categorical sections
Problem:
Current vimrc file is messy, cluttered, and unorganized.

Solution:
Define high-level categories of options and place the settings
accordingly. Moreover, each category is set using a fold so that
when we open the vimrc file, we can see at a high level what is
going on.

Note:
There is future clean-up work to be done with ft and autoload.